isolation level: READ_UNCOMMITTED, READ_COMMITTED, REPEATABLE_READ

+++ src/main/java/com/sweet/n_plus_one_query/dto/projection/OrderDtoProjection.java
@@ -4,18 +4,16 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class OrderDtoProjection {
     String address;
     String productName;
     Long quantity;
-    BigDecimal price;
+    Long price;
     Long orderId;
 
-    public OrderDtoProjection(String address, String productName, Long quantity, BigDecimal price) {
+    public OrderDtoProjection(String address, String productName, Long quantity, Long price) {
         this.address = address;
         this.productName = productName;
         this.quantity = quantity;
